v32sim updates:
  * continuing work on getting VTEX data processed properly
    * globals.h: adding vtex and vsnd members to mem_t
    * load.c: load_memory() now should be loading VTEX data into memory without headers
      * num_vtex and an array of vtex_offset of pertinent values stored in the mem_t for memory page
  * main.c: removing vbinoffset, vtexoffset, and vsndoffset, now that they are in mem_t and doing something
  * ioports.c: fixing typo on fptr type in ioports_set()
  * displaylist.c: updating 'inventory' display to include VTEX information, read from mem_t

The next important step in being able to process VTEX regions is to make sure the data
is loaded into memory correctly; the means with none of the file headers.

The load_memory() function now seems to do that, for V32_BIOS, V32_CART, and V32_MEMC.

I put in an error check, which fails on BIOS. Somehow it thinks the file ends prematurely.
I will need to investigate this. Suspecting a possible off-by-one issue. Disabled the hard
exit for now.

Updating the prompt `inventory` command to include VTEX information, with offsets. That all
is looking like it is storing properly, and some initial debugging of the reported memory
addresses seems to correspond with the correct data.

VSND needs similar treatment, although I have not yet implemented that.

With this, I might be able to start looking into libGD processing of VTEX data. I'd like to
verify all the data has properly been loaded as well.

It is entirely possible now that MEMCards will work: since I made sure to address the file
header structure of MEMCards during my VTEX work on BIOS and CARTs. If there's not a command-
line argument, that's a simple matter of adding.

I think load/unload commands at the prompt already support MEMCards, unless I happened to
leave the pertinent section blank.

+++ ioports.c
@@ -586,7 +586,7 @@ uint8_t  ioports_set (uint16_t  portaddr, int32_t  i32, float  f32, uint8_t  sys
     data_t   *pptr                    = *(ioports+type);           // pointer for sanity
     region_t *rptr                    = NULL;
     int32_t  *iptr                    = NULL;
-    int32_t  *fptr                    = NULL;
+    float    *fptr                    = NULL;
 
     check                             = ioports_chk (portaddr, FLAG_WRITE, sys_force);
     if (check                        == TRUE)
